libpcp: early binding for per-context derived metrics
By calling __dmbind() at the time of definition, we're able to
perform semantic checks synchronously.  This means both syntactic
and semantic errors for per-context derived metrics are checked at
the time of the pmAddDerived*() call.

Note this is not possible for the global derived metrics as their
semantic checks must be delayed until each context has been established
and the metadata is needed.