pmdamssql: add support for Windows authentication
Add a configuration option to mssql.conf and the PMDA
code to allow the 'trusted_connection' pyodbc setting
to be used.  This is a Kerberos-style authentication,
and does not use username/password.  The default PMDA
behaviour is to use the traditional username/password
though, unless trusted=true is set.

During testing (both styles) I've confirmed that this
PMDA does not need to run as root so have enabled the
privilege-dropping mode - it now runs as $PCP_USER.

Resolves Red Hat BZ #1847808