src/pmdas/sample/src/sample.c & qa/335.out: lrand48() is not deterministic
Turns out that the implementation of lrand48() may be different across
platforms, which is a problem for sample.scramble.* and qa/335.  This
was noted on vm37 (OpenBSD 6.7).

Just for this metric, replace lrand48()/srand48() with a local linear
congruential pseudo random number generator that is based on the ANSI C
parameters, and produces _exactly_ the same sequence of pseudo random
numbers on the range (0,2^31-1) on every platform.