qa/1720: increase dbpmda timeout in subtest
qa/1720 was failing (for some) due to a dbpmda timeout from the
qa/statsd/src/cases/15.py script. So add -q 60 to the dbpmda
timeout to avoid that ... and now it gets further but is still
flakey. e.g. if you're editing anything with "pmdastatsd" in it's
filename, the pgrep in qa/statsd/src/utils/pmdastatsd_test_utils.py
might match that when it's looking for valgrind output.

More work needed on this test ... still not passing on my f32 system,
though strangely passing on Nathan's system (hopefully still after this
change).