src/pmlogger/pmlogmv.sh: deal glob and egrep special chars in filenames
Seems like every time I go to use pmlogmv to solve a real problem it
breaks.

This time it was the guards against duplicate oldname matches that
were wrong, and in "fixing" this I exposed a whole Pandora's Box of
Ugliness associated with "special" characters in the oldname and
newname arguments ... in different places and in different ways we
were exposed to egrep(1) regex meta characters (like ., +, [, ..)
and sh(1) glob meta characters (like ?, *, [, ...) and sh(1) special
characters (like ;, |, &, < and >).

This set of issues is now fixed, and qa -g pmlogmv passes (including
a new qa/1272).