qa: assorted remades after pmlogmv change
All are an extra ./ as prefix in oldname from -V, so no change
(this comes from libpcp, not pmlogmv, so I don't want to change it).

These ones test pmlogmv indirectly from pmlogger_daily et al, so
add 'em to the "pmlogmv" group so they won't be missed next time.