src/pmcd/pmdaproc.sh: quieten generation of domain.h and pmns for Python and Perl
Exposed by recent Remove changes, pmdaSetup() runs python and/or perl
to generate these files, but the stderr babble from the PMDA is emitted,
so lots of lines like:

    Info: Initializing, currently in 'notready' state.
    Info: Enabled modules:
    Info: ['execsnoop']
    Info: Configuring modules:
    Info: execsnoop
    Info: Modules configured.
    Info: Initializing modules:
    Info: execsnoop
    Info: execsnoop: Initialized.
    Info: Modules initialized.
    Info: Registering metrics:
    Info: execsnoop
    Info: Metrics registered.
    Info: Registering helpers:
    ...

Syphon this off, and only display it if the generation of the domain.h
or pmns files fails.

If it is needed, the same info will appear in the PMDA's real log
file once it is installed and started, but that does not generally
pollute the .out files of the QA tests.