src/pmlogger/src/gram.y: fix for repeated metric problem
If we were really lucky and had a pmlogger config like:

.... {
	...[A]
	metric-a
	...[B]
	metric-a
}
then metric-a could end up in two fetchctl_t groups, meaning it
would be fetched twice as often as expected, in back-to-back fetches
each time the associated upper-level task was run.

The critical issue is that if the metrics in the [B] group cause
optFetch to split the task's work into two (or more) fetchctl_t
groups, we "lose" track of the first metric-a, and when the second
metric-a is encountered we feed it to optFetch again instead of
skipping it.

This part of the fix ensures that after calling __pmOptFetchAdd() we
then call linkback(tp) to ensure _all_ of the fetchctl_t groups
are linked back to the associated upper level task (tp).

There is also an additional diagnostic here to dump out all of the
fetchctl_t groups after calling __pmOptFetchAdd() when -Doptfetch
and -Ddesperate are in play.