src/pmlogger/src/checks.c: fix for repeated metric problem
In this part of the fix, chk_one() and chk_all() are given
a make over.

chk_one() had a small "same group" logic error (benign I think),
but was missing any useful diagnostics.

chk_all() had the useful diagnostics, but the logic was grossly incomplete
(as in missing) ... this in the code
	/*TODO, not right!*/
should have been a warning!  Arrrgh.
So the routine has been rewritten ... and in so doing now returns 1
correctly for a replicated metric (and no instances specified) in the
same config file logging group, which means we don't call optFetch()
again back in the caller.