qa/1250: refactor to deal with overlapping policies
If we deal with something in the PCP policy modules, and the problem
is later diagnosed as an SELinux problem, the output from audit2allow
can differ.

So filter out all of the expected output so we can focus on the
real differences, and then filter out the ones where the overlap
in policies may lead to a different explanation.