pmlc: add disconnect and sleep commands
A while back we removed the sleep(1) that pmlc used to do after
each disconnect.  There is no functional reason for the sleep and
it was delaying some startup functions.

Fast forward and qa/381 is flakey and failing on some CI platforms
... the problem is that in scripted uses of pmlc (as in qa/381) the
sequence
	connect ...
	...
	connect ...
may fail if the two connect commands are aimed at the same pmlogger
instance (possibly via different but equivalent args), then pmlogger
does not get a chance to cleanup and reset the socket between the
(implicit disconnect before the second connect) and the real connect.

Adding a disconnect command to pmlc allows us to make the disconnect
explicit.

Adding a sleep command allows us a tuning knob to make sure pmlogger
catches up.

So the sequence above can be re-written:
	connect ...
	...
	disconnect
	sleep 100	(msec are the units here)
	connect ...

The new commands do not change the semantics of any existing pmlc
command.

Some existing pmlc tests were remade because of the command summary that
is produced at pmlc startup.