libpcp: wrap avahi_client_new() failure with pmDebugOptions.discovery
pmcd on centos6 (in CI only) was printing messages to stderr
when avah_client_new() fails. The arguments to avah_client_new()
include the AVAHI_CLIENT_NO_FAIL flag, which means the call is
supposed to succeed even if the avahi daemon isn't running, but
not so in CI on centos6 for some reason.

Anyway, this is now wrapped with pmDebugOptions.discovery like
many of the failure messages for other avahi API calls - anyone
relying on avahi will probably be running with that option anyway.

THis should fix the following QA failures in CI on centos6:
051 169 200 243 255 283 462 578 647 648 649 683 695 712 713 714
755 788 805 845 848 855 861 878 964 966 1000 1058 1511.