qa/051.hosts: update hosts used in qa/051
The list was getting quite dated, and I suspect technologies like
Akami, AWS are messing with the test if the DNS does not give the
same IP addr for a single host over a short period of time.

At least I think this why qa/051 is flakey in CI.

The hosts used now all appear to be hosted in ways that mean the
DNS lookups will be deterministic over the short term.