pmdaopenmetrics: add METADATA support in URL configuration files
See description in updated man page pmdaopenmetrics(1). METADATA:
entries in configuration files specifies the metadata for matching
metrics and will over-ride any "PCP" or "PCP5" metadata in the
ingested in-line metric data returned by the URL end-point.

This is intended to be useful for /metrics end-points that do not
include any metadata (or when the applied fallback heuristics are
not correct) and allows the user to specify the exact desired
metadata for the resulting PCP metrics.

A QA test for this is still coming, man page has been updated.

Related: https://github.com/performancecopilot/pcp/issues/799