pmdaopenmetrics: implement 'notes' callbacks for optional labels
Previous work allowed pmdaopenmetrics labels to be tagged as optional
(see 5d056d8) in URL config files but the PM_LABEL_OPTIONAL flag was
not set in the returned labelsets. Now that we have the label_notes
callbacks implemented in libpcp_pmda and its python binding, separate
optional labels in the label callbacks and only return the intrinsic
labels. Optional labels, if any, are now returned in the new notes
callbacks.

Also updated man page. QA test updates still coming for this.

Resolves https://github.com/performancecopilot/pcp/issues/658