src/libpcp/src/logutil.c: refactor __pmGetArchiveEnd_ctx()
GOTO considered helpful.

Rework to deal with customer-reported problem via Nathan.  We were
dereferencing a NULL FILE pointer.

The logic is a bit convoluted here, but it is all to do with the
various ways in which an archive can be corrupted.  The common code
path is unchanged.

Also qa/1280 (new) reproduced the original problem, and qa/417 was
remade because it deals with corrupted archives and now has slightly
different output for one test case.