pmdaelasticsearch: install a default configuration file
Install a standard /etc/pcp-located config file for the
elasticsearch PMDA.  Create a symlink to it from within
the $PCP_PMDA_DIR/elasticsearch location, and in such a
way as to preserve back-compat with an existing es.conf
file (only if that doesn't exist, use the new files).

This puts the config in a persistent location by default
and also makes it easier for automated setup (Ansible).