pmrepconf: new utility to generate pmrep/pcp2xxx configs
When using Ansible to setup automated exporting to other
analysis systems (pcp2elasticsearch, pcp2spark, etc) its
useful to be able to generate a configuration file based
on templates and the system state, just like we do today
for pmlogger with pmlogconf.

The pmrepconf(1) tool shares the pmlogconf configuration
files, but generates pmrep.conf(5) format instead of the
pmlogger(1) format.  The pmlogconf comment-based syntax
for tracking group state is preserved in pmrepconf, but
without the separate per-group interval component.