qa/admin/check-vm: head gasket replacement
Turns out that Debian packaging allows more than one version of
things to be installed, and embeds this week's version number in
the package name, which is a death sentence for the manifest file
and check-vm.

Now that the CI environment is joined at the hip to check-vm, it is
important to get this right ... so allow * in the package name and
root around on the running system to match * with this week's version
number.

And once check-vm understands all of this, check-manifest needs to
watch and learn.