pcp-dstat: fix handling of explicitly requests instances
Steven Daniels reported that dstat used to allow any given
instance (interface, disk, etc) from the command line to be
reported in an individual columns.  The application of inst
culling (cullinsts) was correct for the total column, but
too aggressive in handling command line instances.  This is
fixed, and test qa/1187 updated accordingly.