qa/check-vm: change oil filters and spark plugs, rotate tyres
Fix Nathan's problem ... when a "packaging" command exists on a
strange platform, don't get confused.  In this case it was brew(1)
(not the Mac OS X one) on Fedora 32 where we need to focus on manifest
lines beginning rpm? and not get confused by those starting brew?

Refactor involved shuffling some code around and also cleaning up
the logic a bit.