qa/1116 and 1158: add _notrun when required kernel symbols are missing
For some reason[*], one of my Fedora 32 systems is running a kernel without
blk_account_io_completion (with this kernel 5.8.9-200.fc32.x86_64) and
another (4.18.19-100.fc27.x86_64) has this symbol.

Anyway, if a required kernel symbol is not present for any reason the
test is doomed, so _notrun is the right thing to do.

[*] the reason is probably some grub stickiness that I need to smack with
    a big stick