dbpmda: add getiname command
If "on" then dbpmda will call back into the PMDA to translate internal
instance identifiers from a pmResult into external instance names and
these are reported in the dumped pmResult.

This is visible in the status and help commands, so there are some
QA remakes.

An updated man page and a new qa/1283 are also included.