qa/admin/check-vm: deal with cmp(1) not being there!
For some CI hosts, we run check-vm => packages to install before
cmp(1) is installed, which makes using cmp(1) in check-vm a little
problematic.

We only need a same/not-the-same outcome, so the new _cmp() function
will try the following in order: sum(1) (simpler than cmp(1) if it
is available), cmp(1), diff(1) and if all else fails stat(1) (only
size can be compared with this one, but in the context here if the
files are different, they are going to be different by a lot).