qa/326: use private pmlogger, halve running time
Using the primary pmlogger was causing random failures, I suspect due to
concurrent pmlogger activity (other metrics being logged besides the ones
qa/326 is interested in, causing pmlogger to perform other pmFetch's and
resulting in us missing the state changes).

With a private pmlogger the test is more stable, and the running time
can be reduced.