libpcp_web: fix double-free in binary query/expr functions
series_calculate_order_binary() is used by all the arithmetic
operators for binary expression evaluation, but was incorrectly
freeing and replacing the left operand with a newly allocated
result string. This led to double-free issues when the original
operand strings in the expr tree are free'd. Fix this by passing
in a pointer to the left and right pmSeriesValue structures
from the expression tree, instead of the sds strings directly.

Also clean up some uninitialized stack variables in various places
and assorted other minor changes. There is still quite a bit of
duplicate/repeated code in the expression functions and the long
parameter lists being passed around can probably be consolidated
with node_t pointers instead, so a more substantial refactoring is
coming later, after Nathan merges his memleak fixes.