qa/1282 & qa/admin/other-packages/unavailable: libuv-devel finally understood
For CentOS 8 (and possibly other versions), the libuv-devel package is
being hidden by the module mumbo jumbo.  So depending on whether Mary
the Punter has done:
	$ dnf module install libuv/devel
(yes with a /, not a -) or not, the libuv-devel (yes, with a - not a /)
package is visible to
	$ dnf search
which explains why I was chasing my tail.

It is no longer marked as "unavailable" in the check-vm control file.

And the filter in qa/1282 has been tweaked to remove any babble associated
with libuv-devel.