qa/common.check: change "_service pmlogger stop" semantics for systemd-based systems
This only applies to systems using systemd.

This is a bit of an experiment.  I'm seeing failures in CI that look
like the pmlogger_check.timer is going off in the middle of a test
that has already turned the primary pmlogger off.  This is bad.

So, _service pmlogger (or pcp) stop now => also stop pmlogger_check.timer
_and_ pmlogger_check.service.  They will be automagically restarted when
the matching _service pmlogger (or pcp) start happens, so this change is
asymmetric.