qa/admin/list-packages: creeping replacement for check-vm
Following a suggestion from Andreas, we're moving check-vm and the
associated control files over to a new scheme where we pre-compute
the list of packages needed for a successful PCP build and QA.

From the qa/admin/package-lists README:

There is one file in this directory for each known distro:version:arch
combination e.g. Ubuntu:20.04:x86_64, and the contents of the
file nominate the non-PCP packages that need to be installed for a
successful PCP build and QA (at least as far as CI is concerned).

If the list is the same for all architectures for a given distro:version
combination, then "any" may be used in lieu of a specific architecture.

The format of each file is as follows:

- anything following a hash (#) is assumed to be a comment
- fields are separated by white space
- one line per package starting with the package-name as the first
  field, optionally followed by one or more "tags" (arbitrary test words)

Some examples:

    autoconf
    # mumble
    Spreadsheet::ReadSXC cpan	# don't need to install this

Some standard "tags" and their meaning:

cpan	package needs to be installed with cpan(1)
pip3	package needs to be installed with pip3(1)
notci	do not install this package in CI (usually a container or
	VM conflict)

The ./new script may be used to generate/regenerate a file for the
distro:version:architecture combination on which ./new is run.  This
uses the check-vm infrastructure and is intended to provide a way
to "seed" the file, then any subsequent fine tuning is done by editing
the file in this directory.