qa/admin: expunge all references to perl(...)
The macro perl(foo) is expanded by rpm into the name of the package
that provides the Perl module foo.  Supporting this was adding
no additional value in the check-vm scripts, and increasing the
maintenance load and the chance of errors.

Revert simply using to the proper package names (which were always
there in the manifest) like every other type of package for every
other package manager.