qa/1282: simplify filter
If the (old) check-vm control files are correct (especially the skip
file) then this should suppress false hits for list-packages -c where
a package is one of a number of alternatives that are available, but
we've made a decision to prefer one over all others (and the others
should be in the skip file).

So this commit fixes the skip file for Fedora 32 and removes the
previous additions to the 1282 filter for Fedora 32.