qa/913: rework redis PMDA install/remove logic
Triggered by noticing the redis PMDA was unexpectedly running
after QA was done.

Changed from _prepare_pmda_install() + _restore_pmda_install() to
_prepare_pmda() + _cleanup_pmda().  The latter does a cleaner and
more robust job of restoring the pmcd and PMDA state at the end of
the test to be the same as it was at the start.

We really should make this change throughout and remove the overlapping
functionality in common.check ... but that's a bigger project.