src/pmdas/rpm/Install: add logic to deal with the location of the RPM DB
In current Fedora Rawhide the RPM DB is being converted to sqlite and
the name of the file has changed ... which is why our "notify" logic
has become disconnected.

Optionally add a -r <path> arg to the PMDA if we know where the RPM DB
has moved to, else don't install the PMDA if the RPM DB cannot be found.