src/pmlogctl/pmlogctl.sh: corner case
Exposed by unleashing pmfind, which created a bunch of pmlogger
instances (good), including for the localhost (good), but with the
hostname being `hostname`, not LOCALHOSTNAME (not bad, just not what
pmlogctl was expecting).

Added qa/1763 to exercise this corner case.

Cleaned up a spurious "pmlogctl" in the name of a derived metric used
to implement the conditon() function in a policy file that did not make
sense in the context of pmiectl (which is the pmlogctl script, called
by another name), so qa/1248 and qa/1249 were remade as a consequence.