build/tar: add more missing empty directories
I understand this now.  tar(1) is perfectly able to store empty
directories, but our packing scripts uses the manifest from the build
to provide a list of files (not files+dirs) to tar ... handing the
dirs to tar would not be good because we'd do recursive descent and
potentially package up files not in the manifest.

Created the "checkdirs" script that can be run after Makepkgs to find
directories installed from our makefiles that do not end up in the
packaged tarball or on the "postinst" script ... these are the ones
that need to be added to "postinst".

Also added some systemctl(1) stuff for Arch Linux where we do not
have pacman(1) packaging as yet, and still do tarball packaging.