qa/1601: pmjson failed: Invalid argument ... finally understood
There are a couple of problems here ... the test discards all error
output from curl(1) (I'm not going to try and fix that), then an
empty input is fed to pmjson which emits an error message about as
helpful as ? from ed(1) (I'm not going to try and fix that either).

The root cause is "[sample:2]" in the url passed to curl, which for
some versions of curl produces "bad range in URL".  Changing this to
"%5Bsample%3A2%5D" fixes the issue (actually it is just the [ =>
%5B that's required, but mixing the encoding is only going to make
it less readable).