pmproxy: fix incorrect use of "%u" in redisfmt() for Redis reply.elements
redisReply.elements is size_t, which is a 64bit unsigned int on 64bit
platforms. The "%u" format meta char used by sdscatfmt() in redisfmt()
expects an unsigned int but fails in mysterious ways when passed an
8 byte size_t argument (under the hood in sds, there is a memcpy involved
here and unlike glibc, no compiler assistance with argument size checks).

Hopefully this is the last of the pmproxy<->redis protocol bugs.