libpcp_web: merge Kyoma's "expr_SID" branch, added diags, QA
Merge Kyoma's "expr_SID" branch, supporting /series/query stored
series IDs for canonical expressions and computed descriptors
for functions and arithmetic operators.

Have also added -Dseries and/or -Dquery diagnostics is various
places, notably in redisProcessCallBacks() to print replies as
they are processed and error messages when an unexpected async
disconnect from the redis server occurs (e.g. redis closes the
connection).

For QA - groups pmseries, libpcp_web and pmproxy all passing,
EXCEPT qa/1214 which triggers an assert in redisProcessCallBacks()
where we receive a reply whilst connected but not subscribed or
monitoring. This is still not understood and causes the test to
fail (and pmproxy to abort :P), so it's _notrun for the time being.

Still coming - updates for PMWEBAPI(3) and pmseries(1) and
additional QA (#1093, #1100).

Resolves: https://github.com/performancecopilot/pcp/pull/1114
Resolves: https://github.com/performancecopilot/pcp/issues/1050
Resolves: https://github.com/performancecopilot/pcp/issues/1089