src/pmdas/linux: fix for cases where waitio counters from /proc/stat go backwards
We now understand that the Linux kernel instrumentation does entertain
the possibility that these waitio counters can go backwards (usually
by a small amount).

Keep track of "high-water" values for the all and percpu variants of
waitio from /proc/stat, and use that in lieu of the current value if
the current value is smaller than the "high-water" value.

Based on observation, the deltas are small and the values from
/proc/stat quickly catch up and push the "high-water" mark along.

Big deltas could occur at counter wraps, and the PMDA code preserves
these.