pmieconf: fix default generated config file path, docs + QA
When pmieconf -c is run as root user, the default generated (or
updated) config file is $PCP_VAR_DIR/config/pmie/config.default.
When not run as root the default is $HOME/.pcp/pmie/config.default.

No spec changes needed. Any existing /etc/pcp/pmie/config.default
remains untouched (we don't know if it was intentionally generated
or not so we cant' delete it and we can't replace it with a symlink).
New installs will no longer create it when the pmie service is
first started.

Updated pmieconf(1) man page and added new QA test qa/1688.

Resolves: https://github.com/performancecopilot/pcp/issues/1149