src/libpcp: performance improvements for archives with dynamic indoms
Problem reported by Nathan with this command:
    $ pcp -a 20201109 atop 10
for a standard "all day" PCP archive.  This was taking up to 3 minutes
to get to the first screen update.

These changes introduce restrictions on the number of instances scanned
during interpolation mode processing, based on the timestamped indom
data in the .meta file of the PCP archive.  This reduces the effort for
the "unbound" processing which is particularly difficult for a large
and dynamic indom like 3.9 that underpins the proc.* metrics.

After these changes, the run time for Nathan's case reduces to about
10 seconds.