libpcp_web: handle fabricated SIDs in /series/values RESTAPI calls
Split out common code from pmSeriesQuery() and pmSeriesLoad() to a new
function: series_parse() which we can now use to parse the expression
associated with fabricated SID params passed to /series/values RESTAPI
calls.  series_parse() is declared extern in query.h but not exported.

Update /series/values callback handling when a reply for a SID param
comes back with an array of zero elements - check if the SID is
fabricated and if so, retrieve the persisted expression from Redis.
In the reply callback from that, parse the expression (with series_parse(),
as above) to give an expression tree and time window.

The remaining part (below) is work in progress:

With this we can call a new function similar to series_solve()
to return the time series values and timestamps for the expression
and timing, which can be merged back into the original /series/values
response array.