libpcp_web: series_solve fabricated SIDs in /series/values RESTAPI calls
Following on from 474520025d, after retrieving and parsing the fabricated
SID expression from Redis, create a new seriesQueryBaton with local
callback settings and solve the parsed expression tree via series_solve().
The on_value and on_done callbacks are transfered to the parent baton's
callbacks (which has access to the request and stream), to append to the
response. Note that /series/values accepts a comma separated list of
series parameters, some (or none) of which may be for persistened
expressions, so we hold a reference on the parent baton and release it
in the on_done callback after the series_solve callbacks have completed.

qa/1874 is remade and verifies the /series/values?series=... output
for the "with samples by SID" sub-tests that were previously returning
an empty JSON array.