pmdalinux: network.all metrics for physical interfaces
Allow aggregate metrics (sum) to be formed based on network
interfaces which are detected as physical devices.  A newly
introduced config file (/etc/pcp/linux/interfaces.conf) is
used to provide a regular expression for match all virtual
interface names, which are culled from the 'physical' set,
and all remaining interfaces stats are summed to form the
new metric values.

Resolves https://github.com/performancecopilot/pcp/issues/1159