python api: fix fetchgroup max_insts size with multiple indoms
This resolves a problem observed in pcp-dstat and reported by
Steven Daniels @IBM.  Its not actually a dstat bug though but
rather stems from a logic error in the pmconfig python code -
which is needed to deal with the awkward way the fetchgroup
libpcp code mandates passing a pre-allocated values array.

The bug is in the pmconfig validate_common_options() routine,
dealing with a zero-valued max_insts parameter.  This causes a
dynamic max_inst setting to be calculated, but because it was
testing for zero-valued max_insts *within* the metrics loop it
overwrote max_insts with the first non-zero value it saw.  If
this was a relatively small (cpu) count and next we see a very
much larger indom (disks in this case), we end up using a too-
small max_insts value and ultimately confusion results.

Added a new test 1801 and a dstat-diskfarm archive with a disk
indom containing ~1700 instances - sufficient to reproduce the
bug and exercise the fix.