qa: new helper script: src/qa/src/pmjson_array_sort
This script sorts the JSON { objects } in a single array,
stdin to stdout. Allows most JSON formatted RESTAPI response
data to be included in QA test qualified output, which is
where you need to see it when a pmseries/libpcp_web/pmproxy
test fails.

See qa/1773 for a usage example (in next commit).