libpcp/interp.c: redo the time_caliper() changes
Rather than overloading t_first and t_last, I've opted to use
additional timestamps (t_birth and t_death) to delimit the
times when an instance in known to exist (from the indom data
in the archive).

This leaves the original code less messed with, and more clearly
identifies when we're using the time_caliper() data to trim the
metric-instance searches.