src/pmlogger/src/ports.c: fix permissions mismatch for pmlogger tmp dir
In the GNUmakefile we create and install the directory
$(PCP_TMP_DIR)/pmlogger with the mode 775 (user and group
"pcp").

But in init_ports() we were trying to mkdir2() the same directory
with mode 777, which can lead to this odd QA failure if the (sort
of default) check.callback is enabled and pmlogger is launched at an
unfortunate time:

    *** Failed permissions/ownership checks ***
    /var/lib/pcp/tmp/pmlogger: wrong mode: expected 775 (from src/pmlogger/GNUmakefile), found 777
    drwxrwxrwx 2 pcp pcp 4096 Dec 22 16:05 /var/lib/pcp/tmp/pmlogger
    Fixed.

This commit changes the code in init_ports() to use mode 775.