qa: pmlogger_daily tests need to treat $PCP_LOG_DIR/pmlogger/.NeedRewrite as special
If the file exits at the start of the test (like a QA run just after an
upgrade/install), then move it aside and reinstate it at the end of the
test.

Explains why some pmlogger_daily tests failed from time to time, and
others take a very long time (rewriting all the archives) sometimes.