htop & *curses.h: extend configure to match ProvideCurses.h
Via ProvideCurses.h, pcp-htop has some complicated logic to check
where curses.h or ncurses.h is hiding.

Turns out that on Arch Linux (vm34) this was in a place that our
configure.ac was not checking and we were not setting the correct
HAVE_*CURSES_H macro and the build failed.

So extend the configure.ac checks, and update
src/include/pcp/config.h.in so we're all on the same page.