qa/747: no. cpus fix and knock-on linux PMDA change
qa/747 was not setting $LINUX_NCPUS, so the Linux PMDA then strangely
(it has always been this way), starts with the no. of cpus as returned
by sysconf() and then silently reduces this if the stats files indicate
fewer CPUs are present, but not vice versa.  So the test will pass if
you have enough CPUs on the SUT, but is doomed to fail if the system
running the test has only a few CPUs (like in CI and the QA Farm).

So, for qa/747 we need to grep in proc/stat (from each stats
tarball) to get the right value of $LINUX_NCPUS, then set this in
the environment.

qa/747 still fails (differently now) because of an even stranger
aspect of the Linux PMDA's behaviour ... namely, that if $LINUX_NCPUS
is set then cpu_node_setup() assumes some sort of QA run and maps all
CPUs to node0.  If this conditional mapping is removed from the Linux
PMDA then qa/747 passes.

But now qa/910 fails because it was depending on the strange Linux
PMDA behaviour!

So rework the Linux PMDA to add $LINUX_NNODES to the set of QA knobs
and this one explicity enables the reinstated "map all CPUs to node0"
behaviour (and nothing else).  Added comments in the QA knobs handling
code while I was there to explain what I think is going.

And finally set $LINUX_NNODES in qa/910.