pmdumptext: RH BZ 1132429 clarification
Refer to https://bugzilla.redhat.com/show_bug.cgi?id=1132429

Marco observed "strange" behaviour with pmdumptext, but this was
finally determined to be correct, suggestion some additional man page
love and attention was needed.

Added qa/1844 to also demonstrate a range of permutions of args to
pmdumptext with various metric sources in play and some arguments
without a host: or archive/ prefix.