libpcp_pmda: add new function pmdaCachePurgeCallback()
pmdaCachePurgeCallback is the same as pmdaCachePurge except
it also invokes a callback with the private data pointer
when a cache entry is culled (the callback would typically
free the private data). This is intended for PMDAs that do
not maintain any separate index or references to the private
data, or need to do something else with that data when the
cache entry is culled.

qa/1634 exercises the new function (in later commits).
Man page for PMDACACHE(3) is also updated.