qa/1705: reduce hz, reduce max_throughput
The test was failing on bozo-vm (the new AMD-based one), where hz
values > 1000 are not finding love and comfort.

So change hz from 9999 to 999.

But test5 was still failing because it did not hit the 64KB/sec
output threshold, so lower this to 4KB/sec so we can trigger
the throttle test reliably on slower VMs.