Fix rsyslog pmda to work with modern output
This represents a major update to how rsyslog metrics (`impstats` module)
are handled.

The first prominent change is that the common element in all new metrics
formats is the presence of the `origin=(.*)` value.  This is used to
determine "modern" vs "legacy" format to group the code a bit.

The second prominent change is that we have added metrics for ignored
metric lines emitted by the `impstats` module, unrecognized metric
lines, unrecognized origins, and unmatched origin metric lines.

The third prominent change is to add PCP metrics for all the available
`omelasticsearch` metrics emitted by `impstats`, along with the 2
additional `core.queue` metrics.

The man page for `pmdarsyslog(1)` has been updated to reference the new
"modern" rsyslog configuration format, and the `qa/455` test has been
expanded to cover the supported metrics.