fixes for GH #1140 - PCP_TMPFILE_DIR used in build
$PCP_TMPFILE_DIR does double duty for both build time and run time
(some scripts are used in the build AND installed in the packages).

As outlined in https://github.com/performancecopilot/pcp/issues/1140
if, via configure options, $PCP_TMPFILE_DIR is set to something that
will be a directory after package installation, but does not exist
on the build machine, the build would fail.

Change the files below to check if $PCP_TMPFILE_DIR is set and is
the name of an existing directory before using it, otherwise fallback
to /tmp.

modified:   src/libpcp/src/mk.exports
modified:   src/libpcp/src/mk.pmdbg
modified:   src/pmieconf/xtractnames
modified:   src/pmns/Make.stdpmid