pmdazfs: minor cleanups relating to global variables
Rename all-caps ZFS_PATH to lower-case as its a variable
not a macro.  Initialize it once and only once, and make
it only settable via env (for QA only, not users).  Cull
use of strcpy with no size (fixes Coverity scan issues).

Global variables are guaranteed to be zero-initialized
by the compiler, so some arcstats simplifications made.