qa: rework filtering for pmda.zfs tests using valgrind
These tests were piping all of the output from _run_valgrind into
src/sortinst which was causing a problem if the valgrind command line
was very long, and then the output was tainted with a (cryptic) message:
    Error: input truncated at line 1962. Increase MAXLINELEN.
in the middle of the pminfo output, and there was no valgrind leak
summary report.

The filter now splits the stdout and stderr parts of the _run_valgrind
output, and runs sortinst on the first part, but simply copies the
second part.

To be consistent, changed all these tests: 1530, 1531, 1533 and 1534.

Although the problem really only infected 1530 and 1533 (and these
ones have new .out files in this commit).