src/pmdas/linux_zfs/zfs_pools.c: rework test for directories
readdir() does not guarantee to set d_type, so using this to
test for directories is dangerous.

Turns out d_type _is_ set for /proc, so the production use of the
PMDA is fine, but QA uses "fake" stats files installed in /tmp
or /var/tmp, and herein lies the problem.

Using stat() will work for QA and production.