build: update the redhat.spec to use %files files
One of the major divergences between pcp.spec.in and
redhat.spec was the use of the manifest form the PCP
build vs using the filesystem post-install.  This is
now reconciled and we use the PCP build as the point
of truth.  Over time now I expect we can make these
two specs converge and perhaps eventually be able to
generate redhat.spec from the pcp.spec.in.  There's
still a number of other problem areas though - some
of which will be resolved in followup commits soon.

This solves the immediate "split tmpfiles.d" problem
where we need separate tmpfiles configs for each and
every sub-package, generated and included in %files
sections for each sub-package.