src/pmlc et al: fix botched error checking logic (covscan)
Coverity CID 288181 exposed a problem in addinst() ... the logic here
was really screwed up and has been reimplemented.

Fortunately the error path is only traversed when pmlc is fed a
logging request of the form:
    log ... metric [ instance ... ]
and metric needs to be a non-leaf in the PMNS, and the descendant
leaf metrics in the PMNS are defined over different instance domains.
This is most unlikely to happen in practice.

And indeed this needed extensions to the sample PMDA (an earlier
commit) to be able to exercise a test case (added as qa/1292 and
qa/1293) ... the latter uses valgrind and exposed a false-match that
needed an addition to the valgrind suppressions file for version 3.15.0
(may need this same suppression for other valgrind versions as this
moves through QA).