pmie: fix nasty scheduler issue around pmcd reconnect
There has been a long-standing, and until recently, unidentified
issue causing pmie to apparently "hang" after a pmcd disconnect
and reconnect cycle.

Thanks to some remote debugging by Troy Wilson the problem has been
identified and fixed in this commit.

If pmie had been running for an elapsed time of E (could be measured
in hours, or days, or longer) when the connection to pmcd is lost,
then when the reconnection occurs, rule evaluation is delayed for
E as a result of miscalculating the time when the next rule should
be evaluated after the reconnection.  The problem would have been
seen for each Task (a group of rules requiring metrics from the same
host) associated with the pmcd that went away and came back.

qa/1296 (new) demonstrates the problem in a small test case that
failed on the old pmie and passes once pmie is rebuilt with this
commit.