libpcp: derived metrics: change semantics for delta(x) when x is unsigned
As Andreas observed, making the type of delta(x) the same as the type
of x is broken when x is unsigned.

So this change maps type of x PM_TYPE_U32 -> type of delta() PM_TYPE_64
and type of x PM_TYPE_U64 -> type of delta() PM_TYPE_DOUBLE, so we
can encode the full range of negative and positive values for delta().

qa/1298 (new) tests this using pmval and an archive.

qa/260 and qa/922 were remade to reflect the different metadata for
delta() in the unsigned argument cases.